output models for joint inversion now correctly reflect vp and density used in the inversion
some simplifications and removal of compiler warnings in SurfaceWaveCalculator and SurfaceWaveData

+++ jif3D/SurfaceWaves/SurfaceWaveData.h
@@ -86,7 +86,7 @@ namespace jif3D
           const std::vector<int> &PeriodInd)
         {
           indexmap.clear();
-          for (int ndata = 0; ndata < PairInd.size(); ndata++)
+          for (size_t ndata = 0; ndata < PairInd.size(); ndata++)
             {
               indexmap.insert(
                   std::pair<int, std::tuple<int, int>>(PairInd[ndata],
